fix(tests): widen tile_bytes param to uint64_t (#23 Phase 2c-1) (#27)
In tests/dataflow/test_c_stationary_matmul.cpp, two methods on the
bandwidth-analysis struct were declared:

    uint64_t total_dma_bytes(uint32_t tile_bytes) const;
    uint64_t total_mesh_bytes(uint32_t tile_bytes) const;

but every caller passes the file-local 'uint64_t tile_bytes' from
line 1741 (cfg.tile_bytes() widened on assignment). That mismatch
triggers ~17 MSVC C4244 warnings across the test - every single
uint64_t -> uint32_t entry in the file traces to these two signatures.

Widening both parameters to uint64_t makes the type match the
intended use without changing behaviour (the value is in MB-range
and fits in uint32_t anyway, this is purely a signature fix).

Verified locally on Linux gcc release:
  c_stationary_matmul_test: 14/14 cases, 24/24 assertions

Refs #23

+++ tests/dataflow/test_c_stationary_matmul.cpp
@@ -1569,11 +1569,11 @@ struct BandwidthAnalysis {
         return a_input_tiles + b_input_tiles + c_partial_reads + c_partial_writes + c_output_tiles;
     }
 
-    uint64_t total_dma_bytes(uint32_t tile_bytes) const {
+    uint64_t total_dma_bytes(uint64_t tile_bytes) const {
         return total_dma_tiles() * tile_bytes;
     }
 
-    uint64_t total_mesh_bytes(uint32_t tile_bytes) const {
+    uint64_t total_mesh_bytes(uint64_t tile_bytes) const {
         return mesh_tiles * tile_bytes;
     }
